Sending messages is functional
Sending messages based on an algorithm to determine if the players are active and how many messages they have received per day is implemented. The current limit for receiving messages per player is 3. When a message is sent by a player, it is sent out to 33% or a third of the active player base.